cleaning up and documenting models
 - moving Misc.scala -> misc package
 - Artifact
 - Deprecation
 - Licence
 - Project
