+++ model/src/main/scala/ch.epfl.scala.index.model/Artifact.scala
@@ -1,24 +1,26 @@
 package ch.epfl.scala.index.model
 
-// typelevel/cats-core
+/**
+ * Representation of a scala artifact.
+ *
+ * @param reference simplified artifact name extracted from the artifactId (ex: typelevel/cats-core)
+ * @param releases associated releases with target (scala, scalajs, ...)
+ * @param deprecation optional deprecation tag
+ *
+ * TODO: should we move: name, description, license from release to artifact ?
+ */
 case class Artifact(
-  // simplified artifact name extracted from the artifactId (ex: cats-core)
-  // ex: typelevel/cats-core
   reference: Artifact.Reference,
-  // associated releases with target (scala, scalajs, ...)
   releases: List[Release],
-
   deprecation: Option[Deprecation] = None
-
-  // should we move: name, description, license from release to artifact ?
 )
 
 object Artifact {
+
+  /**
+   * Artifact reference
+   * @param organization the organization name ex: typelevel
+   * @param name the name of this reference ex: cats-core
+   */
   case class Reference(organization: String, name: String)
 }
-
-// see deprecated tag
-case class Deprecation(
-  since: Option[ISO_8601_Date] = None, // ???
-  useInstead: Set[Artifact.Reference] = Set()
-)
