+++ model/src/main/scala/ch.epfl.scala.index.model/Deprecation.scala
@@ -1,0 +1,12 @@
+package ch.epfl.scala.index.model
+
+/**
+ * Deprecation class to mark an artifact as deprecated and point to a bunch of
+ * alternatives instead.
+ * @param since the date since the artifact is deprecated
+ * @param useInstead a list of alternate Artifact references
+ */
+case class Deprecation(
+  since: Option[ISO_8601_Date] = None,
+  useInstead: Set[Artifact.Reference] = Set()
+)
