+++ model/src/main/scala/ch.epfl.scala.index.model/misc/Url.scala
@@ -1,0 +1,7 @@
+package ch.epfl.scala.index.model.misc
+
+/**
+ * Url type - self explained
+ * @param target the url target ex: http://scala-lang.org
+ */
+case class Url(target: String)
