+++ template/src/main/twirl/ch.epfl.scala.index.views/artifact.scala.html
@@ -2,10 +2,10 @@
 @import ch.epfl.scala.index.model.Artifact
 @import ch.epfl.scala.index.model.SemanticVersion
 @import ch.epfl.scala.index.model.Release
-@import ch.epfl.scala.index.model.UserInfo
 @import ch.epfl.scala.index.model.Dependency
 @import ch.epfl.scala.index.model.JavaDependency
 @import ch.epfl.scala.index.model.ScalaDependency
+@import ch.epfl.scala.index.model.misc.UserInfo
 @(project: Project, artifact: Artifact.Reference, version: Option[SemanticVersion], latestStableReleaseOrSelect: List[Release], user: Option[UserInfo])
 
 @staticJs = {
