+++ template/src/main/twirl/ch.epfl.scala.index.views/editproject.scala.html
@@ -1,4 +1,5 @@
 @import ch.epfl.scala.index.model._
+@import ch.epfl.scala.index.model.misc.UserInfo
 @(project: Project, artifact: Artifact.Reference, version: Option[SemanticVersion], user: Option[UserInfo])
 @main(title = s"Edit ${artifact.name}", showSearch = true, searchQuery = "", user) {
   <main id="container-project">
