+++ template/src/main/twirl/ch.epfl.scala.index.views/main.scala.html
@@ -1,4 +1,5 @@
 @import ch.epfl.scala.index.model._
+@import ch.epfl.scala.index.model.misc.UserInfo
 @(title: String, showSearch: Boolean, searchQuery: String, user: Option[UserInfo], staticJs: Option[Html] = None)(content: Html)
 
 <!DOCTYPE HTML>
