+++ template/src/main/twirl/ch.epfl.scala.index.views/notfound.scala.html
@@ -1,4 +1,5 @@
 @import ch.epfl.scala.index.model._
+@import ch.epfl.scala.index.model.misc.UserInfo
 @(user: Option[UserInfo])
 @main(title = "Not found", showSearch = true, searchQuery = "", user) {
 <main>
