+++ template/src/main/twirl/ch.epfl.scala.index.views/paginate.scala.html
@@ -1,4 +1,4 @@
-@import ch.epfl.scala.index.model._
+@import ch.epfl.scala.index.model.misc.Pagination
 @(pagination: Pagination, uri: String)
 
 <nav>
