+++ template/src/main/twirl/ch.epfl.scala.index.views/searchresult.scala.html
@@ -1,4 +1,7 @@
-@import ch.epfl.scala.index.model._
+@import ch.epfl.scala.index.model.Project
+@import ch.epfl.scala.index.model.misc.Pagination
+@import ch.epfl.scala.index.model.misc.UserInfo
+@import ch.epfl.scala.index.model.misc.Url
 
 @(query: String, sorting: Option[String], pagination: Pagination, projects: List[Project], user: Option[UserInfo])
 @main(title = s"Search $query", showSearch = true, searchQuery = "", user) {
